Add files via upload
This is a version of EZR I came up with. It scrapes RhymeZone and appends the data to an excel sheet EZR. Then prints out the words that rhyme with the input from Rhyme zone. It's rough. I am looking to add features from the master branch to it to improve and merge the ideas behind both. ***THIS IS NOT READY FOR THE MASTER BRANCH***